+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All major/notable changes to this project will be documented in this file.
 
+## [1.3.0] - 2025-07-31
+### Added
+- New tab with Grand Prix overview
+  - Round Number
+  - Country
+  - Location
+  - Number of corners
+  - Total track distance
+
 ## [1.2.0] - 2025-07-30
 ### Added
 - Delta time chart to telemetry tab